Removed Jeweled Chalice Rule
After testing trying to get this treasure a few times, I've come to the conclusion that it is *MUCH* easier than I originally thought and is 100% do-able for casuals to do w/ a little less than 5 seconds to spare. B/c of this, I don't think this rule is required at all anymore. (I originally thought that this could result in this rule becoming an option for users to turn on or off, but when it is this easy there really is not much point to it.)

Through my testing of this though, I did discover a glitch where turning in a treasure onto a different pedestal instead of the Safehouse would still count that as completing a check. More details on this bug to come once I know more and can replicate it.