Optimizing Options & Validating said Options
Using Eijebong's AP-Fuzzer tool, I pinpointed the main reasons for this APWorld failing during generation to be both having more bottle bundle items than bundle locations and the max amount of allowed Clockwerk Parts/Keys being as high as 100.

To address this, the "validate_options" method in __init__.py now checks if there would be more bottle bundle items than locations and if so swapping their values. Though I can understand this not being a desired change, until a significant number of locations are added to this APWorld, on a scale of at least 240 or so, this will have to be in place, or at least something similar to it. For example, I could see the argument for either just increasing item bundle size or decreasing location bundle size, but I couldn't decide which would be better to prioritize so I just had them swap values for now.

Also changed the validations involving the number of keys to now decrease the number of keys to goal to match the total rather than increasing the total number of keys to match the required number.

Also adjusted each option related to keys in the Options file to have the max number of keys for each option to be only a max of 50 rather than 100. This number guarantees that generation will ALWAYS succeed.

I would love to eventually increase this back to the max of 100, but we would need an equal number of locations or more added to the game for that to be allowed. In fact, probably the better solution to this would be to check which other location & item related options that significantly affect the total count of either and to decrease the number of each of these key related options accordingly during the validation step. (Currently mainly only the bottle options)

To go along with this, it would be possible to increase the max keys to 80 if we also included the Pickpocketing locations from my branch's pull request. To get us back up to the max of 100 max keys, we would need to put in more locations from somewhere, which I have an idea of what to do about that, but it would involve basically rewriting almost an entire core system of this APWorld for a proper implementation of it. (Adding Job Checkpoints as locations.)

That all being said, again, I can understand not wanting to implement all of these changes, but until this game gets upwards to at least 340 or something more locations/checks to complete, I think either these changes or something similar to them should be implemented. And if we ever do get this amount of locations added to the APWorld, then we can undo more to all of these changes.